requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1

pdfplumber==0.11.0
python-docx==1.1.0

sentence-transformers==2.7.0
torch>=2.2.0

gspread==6.0.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.125.0

openai==1.30.1
python-dotenv==1.0.1